refactor(features): extract form controllers
