+++ src/features/trades/__tests__/useAddTrade.test.tsx
@@ -1,0 +1,43 @@
+import { act, renderHook } from "@testing-library/react-native";
+
+import { createMemoryJsonStorage } from "@/src/services/storage";
+import { createPortfolioStore } from "@/src/store";
+
+import { useAddTrade } from "../useAddTrade";
+
+describe("useAddTrade", () => {
+  it("reviews and confirms a manual buy trade through the feature controller", async () => {
+    const store = createPortfolioStore({ storage: createMemoryJsonStorage() });
+    const { result } = renderHook(() => useAddTrade({ store }));
+
+    act(() => {
+      result.current.setAssetName("Reliance Industries");
+      result.current.setSymbol("RELIANCE");
+      result.current.setTicker("RELIANCE.NS");
+      result.current.setQuantity("2");
+      result.current.setPricePerUnit("100");
+    });
+
+    act(() => {
+      result.current.handleReview();
+    });
+
+    expect(result.current.reviewTrade).toMatchObject({
+      pricePerUnit: 100,
+      quantity: 2,
+      totalValue: 200,
+      type: "buy",
+    });
+
+    await act(async () => {
+      await result.current.handleConfirm();
+    });
+
+    expect(store.getState().assets).toHaveLength(1);
+    expect(store.getState().trades).toHaveLength(1);
+    expect(store.getState().quoteCache[store.getState().assets[0].id]).toMatchObject({
+      price: 100,
+      source: "manual",
+    });
+  });
+});
